fix: widen messages.user_id to BigInteger for large Telegram IDs
Telegram user IDs can exceed INT32 range (e.g. 7435384565), causing
asyncpg DataError on cost insertion. Change column type from Integer
to BigInteger in the model and add an Alembic migration.

+++ bot/db/models.py
@@ -6,7 +6,7 @@ from bot.db.base import Base
 class Message(Base):
     __tablename__ = "messages"
     id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, nullable=False, index=True)
+    user_id = Column(BigInteger, nullable=False, index=True)
     text = Column(Text, nullable=False)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
